Solve building failure when RAY_USE_NEW_GCS=on and remove ERROR_INDEX suffix.

+++ src/common/redis_module/ray_redis_module.cc
@@ -958,8 +958,8 @@ int TableTestAndUpdate_RedisCommand(RedisModuleCtx *ctx,
 
   auto update = flatbuffers::GetRoot<TaskTableTestAndUpdate>(update_buf);
 
-  bool do_update =
-      static_cast<int>(data->scheduling_state()) & update->test_state_bitmask();
+  bool do_update = static_cast<int>(data->scheduling_state()) &
+                   static_cast<int>(update->test_state_bitmask());
 
   if (!is_nil(update->test_scheduler_id()->str())) {
     do_update =
